+++ src/elements/GeoChart/GeoChartScatterPlot/GeoChartScatterPlot.js
@@ -65,6 +65,20 @@ const d3 = (function () {
 const DEFAULT_RADIUS = 1.5
 const DEFAULT_FILL_COLOR = '#69b3a2'
 const DEFAULT_OPACITY = 1
+const CLICKED_STYLE = {
+  stroke: '#9B9B9B',
+  stroke_width: '8',
+  stroke_opacity: '0.4'
+}
+const DEFAULT_STYLE = {
+  stroke: 'white',
+  stroke_width: '2',
+  stroke_opacity: '1'
+}
+const ON_OVER_STYLE = {
+  fill: 'white',
+  stroke_width: '3'
+}
 
 export function render (d3Instance, d3TipInstance, options, globalOptions) {
   const scatterPlotBaseClass = 'geo-chart-scatter-plot-group'
@@ -114,7 +128,7 @@ function renderSingleGroup (group, d3TipInstance, singleGroupOptions, globalOpti
 
   _.each(singleGroupOptions.data, (dot, i) => {
     dot.index = i
-    dot.isClicked = false
+    dot.isClicked = singleGroupOptions.initiallyClickedIndex === i
   })
 
   const radiusScale = d3.scaleSqrt()
@@ -145,27 +159,14 @@ function renderSingleGroup (group, d3TipInstance, singleGroupOptions, globalOpti
     .classed('is-clicked', (d) => classed(d))
     .transition()
     .duration(globalOptions.chart.animationsDurationInMilliseconds)
-    .style('stroke', (d) => stroke(d))
-    .style('stroke-width', (d) => strokeWidth(d))
-    .style('stroke-opacity', (d) => strokeOpacity(d))
+    .attr('style', (d) => applyStyle(d))
     .attr('r', (d) => {
       return singleGroupOptions.groupKey
         ? radiusScale(d[singleGroupOptions.groupKey])
         : _.defaultTo(singleGroupOptions.getRadius(d, d.index), DEFAULT_RADIUS)
     })
-    .style('cursor', () => {
-      return singleGroupOptions.onDotClick
-        ? 'pointer'
-        : 'default'
-    })
-    .style('fill', (d, i) => _.defaultTo(singleGroupOptions.getFillColor(d, d.index), DEFAULT_FILL_COLOR))
     .attr('cx', getCircleCoordinatesFactory('cx'))
     .attr('cy', getCircleCoordinatesFactory('cy'))
-    .style('opacity', (d, i) => {
-      return singleGroupOptions.getOpacity
-        ? singleGroupOptions.getOpacity(d, d.index)
-        : DEFAULT_OPACITY
-    })
 
   dots
     .exit()
@@ -177,29 +178,48 @@ function renderSingleGroup (group, d3TipInstance, singleGroupOptions, globalOpti
 
   setupTooltipEventListeners(allDots, d3TipInstance, singleGroupOptions.tooltip)
 
+  function applyStyle (d) {
+    const color = `fill: ${_.defaultTo(singleGroupOptions.getFillColor(d, d.index), DEFAULT_FILL_COLOR)};`
+    return `${color} ${cursor(d)} ${opacity(d)} ${stroke(d)} ${strokeWidth(d)} ${strokeOpacity(d)}`
+  }
+
   function classed (d) {
-    if (d.index === singleGroupOptions.initiallyClickedIndex) {
-      d.isClicked = true
-    }
     return d.index === singleGroupOptions.initiallyClickedIndex
   }
 
   function stroke (d) {
-    return d.index === singleGroupOptions.initiallyClickedIndex
-      ? '#9B9B9B'
-      : 'white'
+    const value = d.index === singleGroupOptions.initiallyClickedIndex
+      ? CLICKED_STYLE.stroke
+      : DEFAULT_STYLE.stroke
+    return `stroke: ${value};`
   }
 
   function strokeWidth (d) {
-    return d.index === singleGroupOptions.initiallyClickedIndex
-      ? '8'
-      : '2'
+    const value = d.index === singleGroupOptions.initiallyClickedIndex
+      ? CLICKED_STYLE.stroke_width
+      : DEFAULT_STYLE.stroke_width
+    return `stroke-width: ${value};`
   }
 
   function strokeOpacity (d) {
-    return d.index === singleGroupOptions.initiallyClickedIndex
-      ? '0.4'
-      : '1'
+    const value = d.index === singleGroupOptions.initiallyClickedIndex
+      ? CLICKED_STYLE.stroke_opacity
+      : DEFAULT_STYLE.stroke_opacity
+    return `stroke-opacity: ${value};`
+  }
+
+  function opacity (d) {
+    const value = singleGroupOptions.getOpacity
+      ? singleGroupOptions.getOpacity(d, d.index)
+      : DEFAULT_OPACITY
+    return `opacity: ${value};`
+  }
+
+  function cursor (d) {
+    const value = singleGroupOptions.onDotClick
+      ? 'pointer'
+      : 'default'
+    return `cursor: ${value};`
   }
 
   function getSingleDotCSSClasses (d, i) {
@@ -256,23 +276,23 @@ function renderSingleGroup (group, d3TipInstance, singleGroupOptions, globalOpti
   }
 
   function handleMouseOver (d, i) {
-    if (!singleGroupOptions.tooltip) return
+    if (!singleGroupOptions.tooltip && !singleGroupOptions.onDotClick) return
     if (d.isClicked) return
 
     d3.select(this)
-      .style('fill', 'white')
+      .style('fill', ON_OVER_STYLE.fill)
       .style('stroke', (d) => _.defaultTo(singleGroupOptions.getFillColor(d, d.index), DEFAULT_FILL_COLOR))
-      .style('stroke-width', '3')
+      .style('stroke-width', ON_OVER_STYLE.stroke_width)
   }
 
   function handleMouseOut (d, i) {
-    if (!singleGroupOptions.tooltip) return
+    if (!singleGroupOptions.tooltip && !singleGroupOptions.onDotClick) return
     if (d.isClicked) return
 
     d3.select(this)
       .style('fill', (d) => _.defaultTo(singleGroupOptions.getFillColor(d, d.index), DEFAULT_FILL_COLOR))
-      .style('stroke', 'white')
-      .style('stroke-width', '2')
+      .style('stroke', DEFAULT_STYLE.stroke)
+      .style('stroke-width', DEFAULT_STYLE.stroke_width)
   }
 
   function getPreviouslyClickedDot () {
@@ -305,9 +325,9 @@ function renderSingleGroup (group, d3TipInstance, singleGroupOptions, globalOpti
     d.isClicked = true
     element
       .style('fill', color)
-      .style('stroke', '#9B9B9B')
-      .style('stroke-width', '8')
-      .style('stroke-opacity', '0.4')
+      .style('stroke', CLICKED_STYLE.stroke)
+      .style('stroke-width', CLICKED_STYLE.stroke_width)
+      .style('stroke-opacity', CLICKED_STYLE.stroke_opacity)
       .classed('is-clicked', true)
   }
 
@@ -316,9 +336,9 @@ function renderSingleGroup (group, d3TipInstance, singleGroupOptions, globalOpti
     d.isClicked = false
     element
       .style('fill', color)
-      .style('stroke', 'white')
-      .style('stroke-width', '2')
-      .style('stroke-opacity', '1')
+      .style('stroke', DEFAULT_STYLE.stroke)
+      .style('stroke-width', DEFAULT_STYLE.stroke_width)
+      .style('stroke-opacity', DEFAULT_STYLE.stroke_opacity)
       .classed('is-clicked', false)
   }
 }